fixed bugs and naming

+++ api/controllers/friendController.js
@@ -1,17 +1,11 @@
 const { Sequelize } = require('../db/sequelize');
-const { BadRequestError } = require('../errors'); // import NotFoundError and delete BadRequestError
+const { NotFoundError } = require('../errors'); 
 const { Friendship, User } = require('../models');
 const { StatusCodes } = require('http-status-codes');
 const { Op } = require('sequelize');
 
-// searchUsers for the controller name cuz its not clear if its searching among the list of friends or searching users
-const findFriends = async (req, res) => {
-
-  // use query parameter to get the search value
-  // in the url /search?username=jerryl the query parameter 'username' is 'jerryl'
-  // const username = req.query.username
-
-  const query = req.params.id;
+const searchUsers = async (req, res) => {
+  const username = req.params.username;
 
   const searchResult = await User.findAll({
     where: {
@@ -19,7 +13,7 @@ const findFriends = async (req, res) => {
         Sequelize.where(
           Sequelize.fn('LOWER', Sequelize.col('username')),
           {
-            [Sequelize.Op.like]: query.toLowerCase() + '%'
+            [Sequelize.Op.like]: username.toLowerCase() + '%'
           }
         ),
         Sequelize.where(
@@ -32,8 +26,6 @@ const findFriends = async (req, res) => {
     }
   });
 
-  // Just straight away return list of users no need extra attribute name 
-  // json({ searchResult })
   res.status(StatusCodes.OK).json({ users: searchResult });
 };
 
@@ -57,31 +49,28 @@ const getFriends = async (req, res) => {
   const allFriendships = [...requestorFriendships, ...requesteeFriendships];
 
   const friendIds = allFriendships.map(friendship => {
-    if (friendship.requestor === req.user.id) {
-      return friendship.requestee;
-    } else {
-      return friendship.requestor;
-    }
+    return friendship.requestor === req.user.id ? friendship.requestee : friendship.requestor;
   });
 
   const friends = await User.findAll({
     where: {
-      id: friendIds
-    }
+      id: friendIds,
+    },
+    attributes: ['username', 'pfp']
   });
 
   res.status(StatusCodes.OK).json({ friends });
 };
 
 const sendFriendRequest = async (req, res) => {
-  const FriendId = req.params.id; // friendId
+  const friendId = req.params.id;
+  const user = await User.findOne({ where: { id: friendId } });
+  if (!user) {
+    throw new NotFoundError('User does not exist');
+  }
 
-  // validate if the id exists
-  // const user = User.FindOne({ where: { id: friendId } })
-  // if (!user) throw new NotFoundError('User does not exist.')
-
-  Friendship.create({
-    requestee: FriendId,
+  await Friendship.create({
+    requestee: friendId,
     requestor: req.user.id,
     accepted: false
   });
@@ -98,21 +87,17 @@ const getFriendRequests = async (req, res) => {
   });
 
   const requestorIds = friendRequests.map(friendRequest => friendRequest.requestor);
-
-  // Join table first then query the users directly and only need userid, username and pfp attributes
-  const requestors = await User.findAll({
-    where: {
-      id: requestorIds
-    }
-  });
+  const requestors = await User.findAll({ where: { id: requestorIds } });
 
   res.status(StatusCodes.OK).json({ friendRequests: requestors });
 };
 
 const acceptFriendRequest = async (req, res) => {
   const requestorId = req.params.id;
-
-  // validate if the id exists
+  const user = await User.findOne({ where: { id: requestorId } });
+  if (!user) {
+    throw new NotFoundError('User does not exist');
+  }
 
   const friendRequest = await Friendship.findOne({
     where: {
@@ -122,12 +107,13 @@ const acceptFriendRequest = async (req, res) => {
     }
   });
 
-  // if (!friendRequest) throw new NotFoundError('Friend request does not exist.')
+  if (!friendRequest) {
+    throw new NotFoundError('Friend request does not exist.');
+  }
 
   friendRequest.accepted = true;
-  const Requeststatus = await friendRequest.save(); // delete unused variable names so just save() can alr
+  await friendRequest.save(); 
   res.status(StatusCodes.OK).json({ msg: "Friend request accepted successfully." });
-
 };
 
 const declineFriendRequest = async (req, res) => {
@@ -138,26 +124,22 @@ const declineFriendRequest = async (req, res) => {
       requestor: requestorId,
       requestee: req.user.id,
       accepted: false
-    }
+    },
   });
 
-  // do this 3 lines only
-  // if (!friendRequest) throw new NotFoundError('Friend request does not exist.')
-  // await friendRequest.destroy()
-  // res.status(StatusCodes.OK).json({ msg: "Friend request declined successfully." });
-
-  if (friendRequest) {
-    await friendRequest.destroy();
-    res.status(StatusCodes.OK).json({ msg: "Friend request declined successfully." });
-  } else {
-    throw new BadRequestError('Friend request not declined.')
+  if (!friendRequest) {
+    throw new NotFoundError('Friend request does not exist.');
   }
+  await friendRequest.destroy();
+  res.status(StatusCodes.OK).json({ msg: "Friend request declined successfully." });
 };
 
 const removeFriend = async (req, res) => {
   const friendId = req.params.id;
-
-  // validate if the id exists
+  const user = await User.findOne({ where: { id: friendId } });
+  if (!user) {
+    throw new NotFoundError('User does not exist');
+  }
 
   await Friendship.destroy({
     where: {
@@ -167,13 +149,13 @@ const removeFriend = async (req, res) => {
       ]
     }
   });
-  res.status(StatusCodes.OK).json({ msg: "Friend removed successfully." });
 
-  };
+  res.status(StatusCodes.OK).json({ msg: "Friend removed successfully." });
+};
 
 module.exports = {
   getFriends,
-  findFriends,
+  searchUsers,
   sendFriendRequest,
   getFriendRequests,
   acceptFriendRequest,
